tutorial 15
https://www.youtube.com/watch?v=wFV9zPU_Cjg

+++ Vulkan Engine/apps/rotating_triangles_app.cpp
@@ -27,7 +27,7 @@ namespace vk_engine
 
 			const float aspect = renderer.get_aspect_ratio();
 			camera.set_orthographic_projection(-aspect, aspect, -1.f, 1.f, -1.f, 1.f);
-			
+
 			if (const auto command_buffer = renderer.begin_frame())
 			{
 				renderer.begin_swap_chain_render_pass(command_buffer);
@@ -86,7 +86,7 @@ namespace vk_engine
 			auto triangle = vk_game_object::create_game_object();
 			triangle.model = triangle_model;
 			triangle.transform.scale = glm::vec3(.5f, .5f, .5f) + i * 0.025f;
-			triangle.transform.rotation = glm::vec3(0.f, 0.f, glm::radians(45.f)) * glm::vec3(i);
+			triangle.transform.rotation = glm::vec3(0.f, 0.f, glm::radians(45.f)) * glm::vec3(static_cast<float>(i));
 			triangle.color = colors[i % colors.size()];
 
 			game_objects.push_back(std::move(triangle));
